Home Screen Updates, About Page Updates (In Progess), Savings Goal Updates (Remove Goal Button In Progress
Home Screen: I added the logo and changed up the text so it looked a little nicer. Still not sure what to really put here..
About Page: Added our team names but still missing our pictures and a logo.
Savings page: Changed up the layout so it looks a little nicer. Added a button to remove goals but it does NOT WORK. it is just a placeholder.